--- src/module/test/testdt2mysql/test_dt2sql.py
@@ -1,217 +1,0 @@
-# python 3.7.5
-# coding: utf-8
-"""
-dt2sqlモジュール用単体テスト
-実行順番があるので一つのファイルにまとめる
-"""
-import pytest
-import time
-
-from module.dt2sql import (
-    RaceData2sql,
-    ChokuzenData2sql,
-    ResultData2sql,
-    Odds2sql
-)
-from module.getdata import OfficialOdds
-from .common import CommonMethod
-
-WAIT = 0.5
-
-
-@pytest.mark.run(order=4)
-class TestChokuzenInfo2sql(CommonMethod):
-    __target_date = 20200512
-    __jyo_cd = 21
-    __race_no = 1
-    __ci2sql = ChokuzenData2sql()
-
-    @pytest.fixture(scope='class', autouse=True)
-    def insertdata(self):
-        self.__ci2sql.create_table_if_not_exists()
-        self.__ci2sql.insert2table(
-            self.__target_date, self.__jyo_cd, self.__race_no)
-        time.sleep(WAIT)
-
-    cc_col_set = {'race_id', 'datejyo_id',
-                  'temp', 'weather', 'wind_v',
-                  'w_temp', 'wave', 'wind_dr'}
-    cp_col_set = {'waku_id', 'race_id', 'p_name',
-                  'p_weight', 'p_chosei_weight',
-                  'p_tenji_time', 'p_tilt',
-                  'p_tenji_course', 'p_tenji_st'}
-
-    @pytest.mark.parametrize("tb_name, col_set", [
-        ('chokuzen_cond_tb', cc_col_set),
-        ('chokuzen_player_tb', cp_col_set)
-    ])
-    def test_exist_table_raceinfo(self, tb_name, col_set):
-        # カラム名の一致でテスト
-        get_set = super().get_columns2set(tb_name)
-        assert get_set == col_set
-
-    race_id = f"{__target_date}{__jyo_cd:02}{__race_no:02}"
-    cond_col_list = ["race_id", "temp", "weather", "wave"]
-    cond_expected = (
-        int(race_id),
-        24.0,
-        '晴',
-        5
-    )
-    waku_id = f"{__target_date}{__jyo_cd:02}{__race_no:02}1"
-    waku_col_list = ["waku_id", "p_chosei_weight", "p_tenji_time",
-                     "p_tilt", "p_tenji_course", "p_tenji_st"]
-    waku_ex = (
-            int(waku_id),
-            0.0,
-            6.91,
-            -0.5,
-            2,
-            0.11
-        )
-
-    @pytest.mark.parametrize("tb_nm, id_nm, t_id, col_list, expected", [
-        ("chokuzen_cond_tb", "race_id", race_id, cond_col_list, cond_expected),
-        ("chokuzen_player_tb", "waku_id", waku_id, waku_col_list, waku_ex)
-    ])
-    def test_insert2table(self, tb_nm, id_nm, t_id, col_list, expected):
-        res_tpl = super().getdata2tuple(
-            tb_nm,
-            id_nm,
-            t_id,
-            col_list
-        )
-        assert res_tpl == expected
-
-
-@pytest.mark.run(order=5)
-class TestResult2sql(CommonMethod):
-    __target_date = 20200512
-    __jyo_cd = 21
-    __race_no = 1
-    __res2sql = ResultData2sql()
-
-    @pytest.fixture(scope='class', autouse=True)
-    def insertdata(self):
-        self.__res2sql.create_table_if_not_exists()
-        self.__res2sql.insert2table(
-            self.__target_date, self.__jyo_cd, self.__race_no)
-        time.sleep(WAIT)
-
-    rr_col_set = {'race_id', 'datejyo_id', 'temp', 'weather', 'wind_v',
-                  'w_temp', 'wave', 'wind_dr',
-                  'henkantei_list', 'is_henkan', 'kimarite',
-                  'biko', 'payout_3tan', 'popular_3tan',
-                  'payout_3fuku', 'popular_3fuku',
-                  'payout_2tan', 'popular_2tan',
-                  'payout_2fuku', 'popular_2fuku', 'payout_1tan'}
-    rp_col_set = {'waku_id', 'race_id', 'p_rank', 'p_name',
-                  'p_id', 'p_racetime', 'p_course', 'p_st_time'}
-
-    @pytest.mark.parametrize("tb_name, col_set", [
-        ('race_result_tb', rr_col_set),
-        ('p_result_tb', rp_col_set)
-    ])
-    def test_exist_table_raceinfo(self, tb_name, col_set):
-        # カラム名の一致でテスト
-        get_set = super().get_columns2set(tb_name)
-        assert get_set == col_set
-
-    race_id = f"{__target_date}{__jyo_cd:02}{__race_no:02}"
-    race_col_list = ["race_id", "temp", "kimarite", "biko",
-                     "payout_3tan", "popular_3tan"]
-    race_expected = (
-        int(race_id),
-        21.0,
-        '逃げ',
-        '',
-        4000,
-        14
-    )
-    # 5号艇を見る
-    waku_id = f"{__target_date}{__jyo_cd:02}{__race_no:02}5"
-    waku_col_list = ["waku_id", "p_rank", "p_racetime", "p_st_time"]
-    waku_ex = (
-            int(waku_id),
-            -1,
-            -1,
-            0.11
-        )
-
-    @pytest.mark.parametrize("tb_nm, id_nm, t_id, col_list, expected", [
-        ("race_result_tb", "race_id", race_id, race_col_list, race_expected),
-        ("p_result_tb", "waku_id", waku_id, waku_col_list, waku_ex)
-    ])
-    def test_insert2table(self, tb_nm, id_nm, t_id, col_list, expected):
-        res_tpl = super().getdata2tuple(
-            tb_nm,
-            id_nm,
-            t_id,
-            col_list
-        )
-        assert res_tpl == expected
-
-
-@pytest.mark.run(ordre=6)
-class TestOdds2sql(CommonMethod):
-    __target_date = 20200512
-    __jyo_cd = 21
-    __race_no = 1
-
-    # load Official Odds for get keys
-    __ood = OfficialOdds
-
-    @pytest.fixture(scope='class', autouse=True)
-    def insertdata(self):
-        __od2sql = Odds2sql()
-        __od2sql.create_table_if_not_exists()
-        __od2sql.insert2table(
-            self.__target_date, self.__jyo_cd, self.__race_no)
-        time.sleep(WAIT)
-
-    key_set = {'race_id'}
-    three_rentan_key = key_set.union(set(__ood.rentan_keylist(3)))
-    three_renfuku_key = key_set.union(set(__ood.renfuku_keylist(3)))
-    two_rentan_key = key_set.union(set(__ood.rentan_keylist(2)))
-    two_renfuku_key = key_set.union(set(__ood.renfuku_keylist(2)))
-    one_rentan_key = key_set.union(set(__ood.rentan_keylist(1)))
-
-    @pytest.mark.parametrize("tb_name, col_set", [
-        ('odds_3tan_tb', three_rentan_key),
-        ('odds_3fuku_tb', three_renfuku_key),
-        ('odds_2tan_tb', two_rentan_key),
-        ('odds_2fuku_tb', two_renfuku_key),
-        ('odds_1tan_tb', one_rentan_key)
-    ])
-    def test_exist_table_raceinfo(self, tb_name, col_set):
-        # カラム名の一致でテスト
-        get_set = super().get_columns2set(tb_name)
-        assert get_set == col_set
-
-    race_id = int(f"{__target_date}{__jyo_cd:02}{__race_no:02}")
-    three_tan_col_list = ["race_id", "`1-2-3`", "`4-5-6`", "`6-5-4`"]
-    three_tan_expected = (race_id, 31.9, 157.4, 544.4)
-    three_fuku_col_list = ["race_id", "`1-2-3`", "`2-3-4`", "`4-5-6`"]
-    three_fuku_expected = (race_id, 7.6, 24.4, 38.3)
-    two_tan_col_list = ["race_id", "`1-2`", "`4-5`", "`6-5`"]
-    two_tan_expected = (race_id, 10.2, 28.0, 108.3)
-    two_fuku_col_list = ["race_id", "`1-2`", "`2-3`", "`4-5`"]
-    two_fuku_expected = (race_id, 7.9, 14.7, 11.6)
-    one_tan_col_list = ["race_id", "`1`", "`5`"]
-    one_tan_expected = (race_id, 2.7, 1.8)
-
-    @pytest.mark.parametrize("tb_nm, col_list, expected", [
-        ("odds_3tan_tb", three_tan_col_list, three_tan_expected),
-        ("odds_3fuku_tb", three_fuku_col_list, three_fuku_expected),
-        ("odds_2tan_tb", two_tan_col_list, two_tan_expected),
-        ("odds_2fuku_tb", two_fuku_col_list, two_fuku_expected),
-        ("odds_1tan_tb", one_tan_col_list, one_tan_expected),
-    ])
-    def test_insert2table(self, tb_nm, col_list, expected):
-        res_tpl = super().getdata2tuple(
-            tb_nm,
-            "race_id",
-            self.race_id,
-            col_list
-        )
-        assert res_tpl == expected
